+++ calculatrice-ti83-react/src/components/Editors/CatalogViewer.tsx
@@ -37,6 +37,7 @@ const catalogItems = [
   'variance(',
   'xor',
   'R►Pr(', 'R►Pθ(', 'P►Rx(', 'P►Ry(',
+  'Str1', 'Str2', 'Str3', 'Str4', 'Str5', 'Str6', 'Str7', 'Str8', 'Str9',
   '→DMS', '→Dec', '³√(',
   'ΔList(', 'Σ(',
   '∫(',